Move common code to common location
