+++ link-grammar/dict-common/dict-impl.h
@@ -10,3 +10,4 @@ void dictionary_setup_locale(Dictionary dict);
 void dictionary_setup_defines(Dictionary dict);
 void afclass_init(Dictionary dict);
 bool afdict_init(Dictionary dict);
+void affix_list_add(Dictionary afdict, Afdict_class *, const char *);
