+++ link-grammar/dict-file/dictionary.c
@@ -51,52 +51,6 @@ static inline char * deinflect(const char * str)
 	return s;
 }
 
-/* The affix dictionary is represented as a dynamically allocated array with
- * an element for each class (connector type) in the affix file. Each element
- * has a pointer to an array of strings which are the punctuation/affix
- * names. */
-
-const char * afdict_classname[] = { AFDICT_CLASSNAMES };
-
-/**
- * Find the affix table entry for given connector name.
- * If the connector name is not in the table, return NULL.
- */
-Afdict_class * afdict_find(Dictionary afdict, const char * con, bool notify_err)
-{
-	const char ** ac;
-
-	for (ac = afdict_classname;
-	     ac < &afdict_classname[ARRAY_SIZE(afdict_classname)]; ac++)
-	{
-		if (0 == strcmp(*ac, con))
-			return &afdict->afdict_class[ac - afdict_classname];
-	}
-	if (notify_err) {
-		prt_error("Warning: Unknown class name %s found near line %d of %s.\n"
-		          "\tThis class name will be ignored.\n",
-		          con, afdict->line_number, afdict->name);
-	}
-	return NULL;
-}
-
-#define AFFIX_COUNT_MEM_INCREMENT 64
-
-static void affix_list_add(Dictionary afdict, Afdict_class * ac,
-		const char * affix)
-{
-	if (NULL == ac)  return; /* ignore unknown class name */
-	if (ac->mem_elems <= ac->length)
-	{
-		size_t new_sz;
-		ac->mem_elems += AFFIX_COUNT_MEM_INCREMENT;
-		new_sz = ac->mem_elems * sizeof(const char *);
-		ac->string = (char const **) realloc((void *)ac->string, new_sz);
-	}
-	ac->string[ac->length] = string_set_add(affix, afdict->string_set);
-	ac->length++;
-}
-
 static void load_affix(Dictionary afdict, Dict_node *dn, int l)
 {
 	Dict_node * dnx = NULL;
@@ -140,255 +94,6 @@ static void load_affix(Dictionary afdict, Dict_node *dn, int l)
 		dnx = dn->left;
 		xfree((char *)dn, sizeof(Dict_node));
 	}
-}
-
-#ifdef AFDICT_ORDER_NOT_PRESERVED
-static int revcmplen(const void *a, const void *b)
-{
-	return strlen(*(char * const *)b) - strlen(*(char * const *)a);
-}
-#endif /* AFDICT_ORDER_NOT_PRESERVED */
-
-/**
- * Traverse the main dict in dictionary order, and extract all the suffixes
- * and prefixes - every time we see a new suffix/prefix (the previous one is
- * remembered by w_last), we save it in the corresponding affix-class list.
- * The saved affixes don't include the infix mark.
- */
-static void get_dict_affixes(Dictionary dict, Dict_node * dn,
-                             char infix_mark, char * w_last)
-{
-	const char *w;         /* current dict word */
-	const char *w_sm;      /* SUBSCRIPT_MARK position in the dict word */
-	size_t w_len;          /* length of the dict word */
-	Dictionary afdict = dict->affix_table;
-
-	if (dn == NULL) return;
-	get_dict_affixes(dict, dn->right, infix_mark, w_last);
-
-	w = dn->string;
-	w_sm = strrchr(w, SUBSCRIPT_MARK);
-	w_len = (NULL == w_sm) ? strlen(w) : (size_t)(w_sm - w);
-	if (w_len > MAX_WORD)
-	{
-		prt_error("Error: word '%s' too long (%zd), program may malfunction\n",
-		          w, w_len);
-		w_len = MAX_WORD;
-	}
-	/* (strlen(w_last) can be cached for speedup) */
-	if ((strlen(w_last) != w_len) || (0 != strncmp(w_last, w, w_len)))
-	{
-		strncpy(w_last, w, w_len);
-		w_last[w_len] = '\0';
-
-		if (infix_mark == w_last[0])
-		{
-			affix_list_add(afdict, &afdict->afdict_class[AFDICT_SUF], w_last+1);
-		}
-		else
-		if (infix_mark == w_last[w_len-1])
-		{
-			w_last[w_len-1] = '\0';
-			affix_list_add(afdict, &afdict->afdict_class[AFDICT_PRE], w_last);
-			w_last[w_len-1] = infix_mark;
-		}
-	}
-
-	get_dict_affixes(dict, dn->left, infix_mark, w_last);
-}
-
-/**
- * Concatenate the definitions for the given affix class.
- * This allows specifying the characters in different definitions
- * instead in a one long string, e.g. instead of:
- * ""«»《》【】『』`„": QUOTES+;
- * One can specify (note the added spaces):
- * """  «»  《》 【】 『』  ` „: QUOTES+;
- * Or even:
- * """: QUOTES+;
- * «» : QUOTES+;
- * etc.
- * Note that if there are no definitions or only one definition, there is
- * nothing to do.
- * The result is written to the first entry.
- * @param classno The given affix class.
- */
-static void concat_class(Dictionary afdict, int classno)
-{
-	Afdict_class * ac;
-	size_t i;
-	dyn_str * qs;
-
-	ac = AFCLASS(afdict, classno);
-	if (1 >= ac->length) return;
-
-	qs = dyn_str_new();
-	for (i = 0; i < ac->length; i++)
-		dyn_strcat(qs, ac->string[i]);
-
-	ac->string[0] = string_set_add(qs->str, afdict->string_set);
-	dyn_str_delete(qs);
-}
-
-/* Compare lengths of strings, for qsort */
-static int cmplen(const void *a, const void *b)
-{
-	const char * const *sa = a;
-	const char * const *sb = b;
-	return strlen(*sb) - strlen(*sa);
-}
-
-/**
- * Initialize several classes.
- * In case of a future dynamic change of the affix table, this function needs to
- * be invoked again after the affix table is re-constructed (changes may be
- * needed - especially to first free memory and initialize the affix dict
- * structure.).
- */
-#define D_AI 11
-bool afdict_init(Dictionary dict)
-{
-	Afdict_class * ac;
-	Dictionary afdict = dict->affix_table;
-
-	/* FIXME: read_entry() builds word lists in reverse order (can we
-	 * just create the list top-down without breaking anything?). Unless
-	 * it is fixed to preserve the order, reverse here the word list for
-	 * each affix class. */
-	for (ac = afdict->afdict_class;
-		  ac < &afdict->afdict_class[ARRAY_SIZE(afdict_classname)]; ac++)
-	{
-		int i;
-		int l = ac->length - 1;
-		const char * t;
-
-		for (i = 0;  i < l; i++, l--)
-		{
-			t = ac->string[i];
-			ac->string[i] = ac->string[l];
-			ac->string[l] = t;
-		}
-	}
-
-	/* Create the affix lists */
-	ac = AFCLASS(afdict, AFDICT_INFIXMARK);
-	if ((1 < ac->length) || ((1 == ac->length) && (1 != strlen(ac->string[0]))))
-	{
-		prt_error("Error: afdict_init: Invalid value for class %s in file %s"
-		          " (should have been one ASCII punctuation - ignored)\n",
-		          afdict_classname[AFDICT_INFIXMARK], afdict->name);
-		free((void *)ac->string);
-		ac->length = 0;
-		ac->mem_elems = 0;
-		ac->string = NULL;
-	}
-	/* XXX For now there is a possibility to use predefined SUF and PRE lists.
-	 * So if SUF or PRE are defined, don't extract any of them from the dict. */
-	if (1 == ac->length)
-	{
-		if ((0 == AFCLASS(afdict, AFDICT_PRE)->length) &&
-		    (0 == AFCLASS(afdict, AFDICT_SUF)->length))
-		{
-			char last_entry[MAX_WORD+1] = "";
-			get_dict_affixes(dict, dict->root, ac->string[0][0], last_entry);
-		}
-	}
-	else
-	{
-		/* No INFIX_MARK - create a dummy one that always mismatches */
-		affix_list_add(afdict, &afdict->afdict_class[AFDICT_INFIXMARK], "");
-	}
-
-	if (verbosity_level(+D_AI))
-	{
-		size_t l;
-
-		for (ac = afdict->afdict_class;
-		     ac < &afdict->afdict_class[ARRAY_SIZE(afdict_classname)]; ac++)
-		{
-				if (0 == ac->length) continue;
-				lgdebug(+0, "Class %s, %zd items:",
-				        afdict_classname[ac-afdict->afdict_class], ac->length);
-				for (l = 0; l < ac->length; l++)
-					lgdebug(0, " '%s'", ac->string[l]);
-				lgdebug(0, "\n");
-		}
-	}
-#undef D_AI
-
-	/* Store the SANEMORPHISM regex in the unused (up to now)
-	 * regex_root element of the affix dictionary, and precompile it */
-	assert(NULL == afdict->regex_root, "SM regex is already assigned");
-	ac = AFCLASS(afdict, AFDICT_SANEMORPHISM);
-	if (0 != ac->length)
-	{
-		int rc;
-
-		Regex_node *sm_re = malloc(sizeof(*sm_re));
-		dyn_str *rebuf = dyn_str_new();
-
-		/* The regex used to be converted to: ^((original-regex)b)+$
-		 * In the initial wordgraph version word boundaries are not supported,
-		 * so instead it is converted to: ^(original-regex)+$ */
-#ifdef WORD_BOUNDARIES
-		dyn_strcat(rebuf, "^((");
-#else
-		dyn_strcat(rebuf, "^(");
-#endif
-		dyn_strcat(rebuf, ac->string[0]);
-#ifdef WORD_BOUNDARIES
-		dyn_strcat(rebuf, ")b)+$");
-#else
-		dyn_strcat(rebuf, ")+$");
-#endif
-		sm_re->pattern = strdup(rebuf->str);
-		dyn_str_delete(rebuf);
-
-		afdict->regex_root = sm_re;
-		sm_re->name = strdup(afdict_classname[AFDICT_SANEMORPHISM]);
-		sm_re->re = NULL;
-		sm_re->next = NULL;
-		sm_re->neg = false;
-		rc = compile_regexs(afdict->regex_root, afdict);
-		if (rc) {
-			prt_error("Error: afdict_init: Failed to compile "
-			          "regex '%s' in file %s, return code %d\n",
-			          afdict_classname[AFDICT_SANEMORPHISM], afdict->name, rc);
-			return false;
-		}
-		lgdebug(+5, "%s regex %s\n",
-		        afdict_classname[AFDICT_SANEMORPHISM], sm_re->pattern);
-	}
-
-	/* sort the UNITS list */
-	/* Longer unit names must get split off before shorter ones.
-	 * This prevents single-letter splits from screwing things
-	 * up. e.g. split 7gram before 7am before 7m
-	 */
-	ac = AFCLASS(afdict, AFDICT_UNITS);
-	if (0 < ac->length)
-	{
-		qsort(ac->string, ac->length, sizeof(char *), cmplen);
-	}
-
-#ifdef AFDICT_ORDER_NOT_PRESERVED
-	/* pre-sort the MPRE list */
-	ac = AFCLASS(afdict, AFDICT_MPRE);
-	if (0 < ac->length)
-	{
-		/* Longer subwords have priority over shorter ones,
-		 * reverse-sort by length.
-		 * XXX mprefix_split() for Hebrew depends on that. */
-		qsort(ac->string, ac->length, sizeof(char *), revcmplen);
-	}
-#endif /* AFDICT_ORDER_NOT_PRESERVED */
-
-	concat_class(afdict, AFDICT_QUOTES);
-	concat_class(afdict, AFDICT_BULLETS);
-
-	if (! anysplit_init(afdict)) return false;
-
-	return true;
 }
 
 static void free_llist(Dictionary dict, Dict_node *llist)
@@ -497,6 +202,9 @@ dictionary_six_str(const char * lang,
 		goto failure;
 	}
 	if (! afdict_init(dict))
+		goto failure;
+
+	if (! anysplit_init(dict->affix_table))
 		goto failure;
 
 	/*
